+++ install_deps.R
@@ -1,0 +1,5 @@
+if (!requireNamespace("BiocManager", quietly = TRUE))
+    install.packages("BiocManager")
+
+BiocManager::install("BiocParallel")
+BiocManager::install("Biostrings")