+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
 library(minSNPs)
+library(Biostrings)
 
 test_check("minSNPs")